fix product controller

+++ backend/controller/Products.js
@@ -1,5 +1,6 @@
 import Product from "../models/ProductModel.js";
 import User from "../models/UserModel.js";
+import { Op } from "sequelize";
 
 export const getProducts = async (req, res) => {
 
@@ -7,16 +8,20 @@ export const getProducts = async (req, res) => {
         let response;
         if(req.role === 'admin'){
             response = await Product.findAll({
+                attributes: ['uuid', 'name', 'price'],
                 include: [{
                     model: User,
+                    attributes: ['name', 'email']
                 }]
             });
 
         }else{
             response = await Product.findAll({
+                attributes: ['uuid', 'name', 'price'],
                 where: {UserId: req.UserId},
                 include: [{
                     model: User,
+                    attributes: ['name', 'email']
                 }]
             });
         }
@@ -33,19 +38,44 @@ export const getProducts = async (req, res) => {
 export const getProductById = async (req, res) => {
 
     try {
-        const response = await Product.findOne({
-            where: {id: req.params.id}
+
+        const product = await Product.findOne({
+            where: {uuid: req.params.id},
         });
+
+        if(!product) return res.status(404).json({message: "Product not found"});
+
+        let response;
+        if(req.role === 'admin'){
+            response = await Product.findOne({
+                attributes: ['uuid', 'name', 'price'],
+                where: {id: product.id},
+                include: [{
+                    model: User,
+                    attributes: ['name', 'email']
+                }]
+            });
+
+        }else{
+            response = await Product.findOne({
+                attributes: ['uuid', 'name', 'price'],
+                where: {
+                    [Op.and]: [{id: product.id}, {UserId: req.UserId}],
+                },
+                include: [{
+                    model: User,
+                    attributes: ['name', 'email']
+                }]
+            });
+        }
 
         res.status(200).json(response);
        
     } catch (error) {
-            
-            res.status(500).json(error);
-        }
-        
-  
-    
+
+        res.status(500).json(error);
+    }
+
 }
 
 export const createProduct = async (req, res) => {
@@ -64,27 +94,43 @@ export const createProduct = async (req, res) => {
     } catch (error) {
         res.status(500).json({message: "Something went wrong"});
     }
-
-
 }
 
 export const updateProduct = async (req, res) => {
 
-    const {name, price} = req.body;
-    const {id} = req.params;
-
     try {
-        await Product.update({
-            name: name,
-            price: price,
-        },
-        {
-            where: {id: id}
+
+        const product = await Product.findOne({
+            where: {uuid: req.params.id},
         });
 
-        res.status(200).json({message: "Product updated successfully"});
+        if(!product) return res.status(404).json({message: "Product not found"});
 
+        const {name, price} = req.body;
+        if(req.role === 'admin'){
+            await Product.update({
+                name: name,
+                price: price
+            },{
+                where: {id: product.id}
+            });
+        }else{
+            if(product.UserId !== req.UserId) return res.status(403).json({message: "Unauthorized"});
+            await Product.update({
+                name: name,
+                price: price
+            },{
+                where: {
+                    [Op.and]: [{id: product.id}, {UserId: req.UserId}],
+                }
+            });
+           
+        }
+
+        res.status(200).json({message: "Product updated successfully"});
+       
     } catch (error) {
+
         res.status(500).json({message: "Something went wrong"});
     }
 
@@ -92,16 +138,31 @@ export const updateProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
 
-    const {id} = req.params;
-
     try {
-        await Product.destroy({
-            where: {id: id}
+
+        const product = await Product.findOne({
+            where: {uuid: req.params.id},
         });
 
-        res.status(200).json({message: "Product deleted successfully"});
+        if(!product) return res.status(404).json({message: "Product not found"});
 
+        if(req.role === 'admin'){
+            await Product.destroy({
+                where: {id: product.id}
+            });
+        }else{
+            if(product.UserId !== req.UserId) return res.status(403).json({message: "Unauthorized"});
+            await Product.destroy({
+                where: {
+                    [Op.and]: [{id: product.id}, {UserId: req.UserId}],
+                }
+            });
+        }
+
+        res.status(200).json({message: "Product deleted successfully"});
+       
     } catch (error) {
+
         res.status(500).json({message: "Something went wrong"});
     }
